requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
python-multipart>=0.0.18
pydantic>=2.10,<3
pillow>=11.0,<12
torch>=2.5,<3
transformers>=4.47,<5
accelerate>=1.2,<2
torchaudio>=2.5,<3
soundfile>=0.13,<1
jinja2>=3.1,<4
aiofiles>=24.1,<25
